% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_filter_data.R
\name{categorize_data}
\alias{categorize_data}
\title{categorize_data}
\usage{
categorize_data(data)
}
\arguments{
\item{data}{The data you want to categorize.}
}
\value{
Returns your data. But categorized.
}
\description{
Categorizes your data (Marktwert, Platzierung and Punkte) from continuous/discrete values to categoric values.
}
\examples{
categorize_data(data)
}
