% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_graphic_functions.R
\name{frame_performance}
\alias{frame_performance}
\title{frame_performance}
\usage{
frame_performance(data, performance)
}
\arguments{
\item{data}{The data you want to be rated.}

\item{performance}{The performances you´re looking for. Put in qn element of {-2,-1,0,1,2}.}
}
\value{
Returns a dataframe.
}
\description{
Gives you a frame, which only contains really good (2), good (1), normal (0), bad (-1), really bad (-1) team-performances.
}
\examples{
frame_performance(data,2) returns a frame consisting out of teams with low marketvalue and high placement.
}
