% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_filter_data.R
\name{get_lastfullseason}
\alias{get_lastfullseason}
\title{get_lastfullseason}
\usage{
get_lastfullseason()
}
\value{
Returns you an integer.
}
\description{
Gets you the last fully completed season.
}
