% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PythonWrapper.R
\name{load_data_from_website}
\alias{load_data_from_website}
\title{Title}
\usage{
load_data_from_website(
  saison_von,
  saison_bis,
  liga,
  output = "data",
  append = TRUE
)
}
\arguments{
\item{saison_von}{}

\item{saison_bis}{}

\item{liga}{}

\item{output}{}

\item{append}{}
}
\description{
Title
}
