% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_graphic_functions.R
\name{plot_data_placements}
\alias{plot_data_placements}
\title{plot_data_placements}
\usage{
plot_data_placements(data)
}
\arguments{
\item{data}{The data you want to plot from.}
}
\value{
Returns a plot.
}
\description{
Plots marketvalue and placements of your data
}
\examples{
plot_data_placements(data)
}
