% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_graphic_functions.R
\name{plot_data_points}
\alias{plot_data_points}
\title{plot_data_points}
\usage{
plot_data_points(data)
}
\arguments{
\item{data}{The data you want to plot from.}
}
\value{
Returns a plot.
}
\description{
Plots marketvalue and points of your data.
}
\examples{
plot_data_points(data)
}
