% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_graphic_functions.R
\name{plot_last10years}
\alias{plot_last10years}
\title{plot_last10years}
\usage{
plot_last10years(data, team, categoric = TRUE)
}
\arguments{
\item{data}{The data, which you want to use.}

\item{team}{The team you want to get the performance from.}

\item{categoric}{If true you plot in categoric variables. If false you plot continuous.}
}
\value{
Returns a plot.
}
\description{
Plots the performance of one team over the last 10 years.
}
\examples{
plot_last10years(data,"FC Schalke 04",categoric=TRUE)
}
