% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numeric_graphic_functions.R
\name{plot_oneleague}
\alias{plot_oneleague}
\title{plot_oneleague}
\usage{
plot_oneleague(data, liga, saison = get_lastfullseason())
}
\arguments{
\item{data}{The data you want to plot.}

\item{liga}{The league you want to plot.}

\item{saison}{The season you want to plot.}
}
\value{
Returns a plot.
}
\description{
Plots marketvalue and placements of  one single season from the choosen league.
The red lines symbolize the categoric classification of "Marktwert" and "Platzierung".
}
\examples{
plot_oneleague(data,"Premier League", 2018)
}
