% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_and_filter_data.R
\name{read_data_from_csv}
\alias{read_data_from_csv}
\title{read_data_from_csv}
\usage{
read_data_from_csv(relPath)
}
\arguments{
\item{relPath}{The sourcepath of your CSV.}
}
\value{
Returns your data.
}
\description{
Generates your data out of a CSV.
}
\examples{
read_data_from_csv("< your cs-path >")
}
