% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chi-Squared-Test.R
\name{chi_squared_test}
\alias{chi_squared_test}
\title{Title}
\usage{
chi_squared_test(data, method, reps)
}
\arguments{
\item{data}{}

\item{method}{}

\item{reps}{}
}
\value{
No return value
}
\description{
Title
}
\examples{
chi_squared_test(bigFive,"theory",1000)
}
